requirements.txt
telethon
python-dotenv
pandas
pillow
psycopg2-binary
opencv-python
torch 
torchvision


